+++ NEWS.md
@@ -4,6 +4,7 @@
 * Fixed bug where `days_week = c(1, 7)` was output to `"0/6"` instead of `"0,6"`
 * Document internal functions
 * Test `x/y` edge cases (#9)
+* Basic test coverage to 100% (#8)
 
 # dialga 0.0.0.9001